Changes of version: 20070721

+++ libewf/libewf_read.h
@@ -56,7 +56,7 @@
 extern "C" {
 #endif
 
-int8_t libewf_read_build_index( LIBEWF_INTERNAL_HANDLE *internal_handle );
+int libewf_read_build_index( LIBEWF_INTERNAL_HANDLE *internal_handle );
 EWF_SECTION *libewf_read_sections_from_segment( LIBEWF_INTERNAL_HANDLE *internal_handle, uint16_t segment_number );
 
 ssize_t libewf_read_chunk( LIBEWF_INTERNAL_HANDLE *internal_handle, uint32_t chunk, uint32_t chunk_offset, void *buffer, size_t size );
